metricas de log arregladas

+++ kernel/src/pcb.c
@@ -165,11 +165,11 @@ void log_metricas_estado(t_pcb* pcb){
     if (pcb == NULL) return;
 
     log_info(logger, "%d - Métricas de estado:", pcb->pid);
-    log_info(logger, "NEW (%d) (%d)", pcb->metricas_estado[NEW], pcb->metricas_tiempo[NEW]);
-    log_info(logger, "READY (%d) (%d)", pcb->metricas_estado[READY], pcb->metricas_tiempo[READY]);
-    log_info(logger, "EXEC (%d) (%d)", pcb->metricas_estado[EXEC], pcb->metricas_tiempo[EXEC]);
-    log_info(logger, "BLOCKED (%d) (%d)", pcb->metricas_estado[BLOCKED], pcb->metricas_tiempo[BLOCKED]);
-    log_info(logger, "SUSP_BLOCKED (%d) (%d)", pcb->metricas_estado[SUSP_BLOCKED], pcb->metricas_tiempo[SUSP_BLOCKED]);
-    log_info(logger, "SUSP_READY (%d) (%d)", pcb->metricas_estado[SUSP_READY], pcb->metricas_tiempo[SUSP_READY]);
-    log_info(logger, "EXIT (%d) (%d)", pcb->metricas_estado[EXIT], pcb->metricas_tiempo[EXIT]);
+    log_info(logger, "NEW (%d) (%d)", pcb->metricas_estado[NEW], (int)pcb->metricas_tiempo[NEW]);
+    log_info(logger, "READY (%d) (%d)", pcb->metricas_estado[READY], (int)pcb->metricas_tiempo[READY]);
+    log_info(logger, "EXEC (%d) (%d)", pcb->metricas_estado[EXEC], (int)pcb->metricas_tiempo[EXEC]);
+    log_info(logger, "BLOCKED (%d) (%d)", pcb->metricas_estado[BLOCKED], (int)pcb->metricas_tiempo[BLOCKED]);
+    log_info(logger, "SUSP_BLOCKED (%d) (%d)", pcb->metricas_estado[SUSP_BLOCKED], (int)pcb->metricas_tiempo[SUSP_BLOCKED]);
+    log_info(logger, "SUSP_READY (%d) (%d)", pcb->metricas_estado[SUSP_READY], (int)pcb->metricas_tiempo[SUSP_READY]);
+    log_info(logger, "EXIT (%d) (%d)", pcb->metricas_estado[EXIT], (int)pcb->metricas_tiempo[EXIT]);
 }
